requirements.txt
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
requests>=2.31
python-dotenv>=1.0
rapidfuzz>=3.0
